# bpo-26828: Add __length_hint__() to builtin map iterator