--- a/Misc/NEWS.d/next/Core and Builtins/2019-06-27-16-18-00.bpo-37435.LUdu4p.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2019-06-27-16-18-00.bpo-37435.LUdu4p.rst
@@ -1,0 +1,2 @@
+The iterator provided by builtin :func:`map` now provides a `__length_hint__`
+method.  Patch by Nicholas Musolino.