@nmusolino 2019-06-27
Thank you, @vstinner , for your comment.  I can run some ad-hoc benchmarks with timeit, and also try running `pyperformance`.

It is conceivable that this change would significantly improve performance for large sizes, but make performance worse for very short iterators (under 8 elements), since more work is being done to compute the length hint, with no (re)allocation savings.