@vstinner 2019-06-27
> Thank you, @vstinner , for your comment. I can run some ad-hoc benchmarks with timeit, and also try running pyperformance.

Micro-benchmarks using pyperf time would be acceptable for me ;-) (I don't expect any significant impact on pyperformance benchmarks.)

> It is conceivable that this change would significantly improve performance for large sizes, but make performance worse for very short iterators (under 8 elements), since more work is being done to compute the length hint, with no (re)allocation savings.

Well, https://bugs.python.org/issue26828 has been closed...