@nmusolino 2019-06-29
Here are the results of a microbenchmark as suggested by @vstinner.  It took me some time to get a PGO/LTO build working.

master:  commit 97d15b1ee0
with-length-hint:  commit d6d90d1d3b

Both were built with:
```
mkdir build_$(git log -n1 --format="%h")
cd build_$(git log -n1 --format="%h")
../configure --enable-optimizations --with-lto --with-openssl=$(brew --prefix openssl)
emacs -nw Makefile  # Add `-x test_os` to exclude test_os from PGO profile generation run
make -s -j20 profile-opt
```

System specs:  iMac with 3.8 GHz Intel Core i5, 16 GB DDR4 memory, mac OS 10.13.6

Profiled the following expression with `timeit`:  `list(map(lambda x:x, range(n)))`
```
# profiled expression:   
for i in 0 1 5 10 20 25 50 100 250 500 1000 10_000 100_000 1_000_000 5_000_000 10_000_000 50_000_000; do
    printf "%10s %24s %24s\n" $i \
        "$(./build_97d15b1ee0/python.exe -m timeit "list(map(lambda x:x, range($i)))" | cut -f2 -d':')" \
        "$(./build_d6d90d1d3b/python.exe -m timeit "list(map(lambda x:x, range($i)))" | cut -f2 -d':')";
done
```

```
         0        398 nsec per loop        367 nsec per loop
         1        446 nsec per loop        486 nsec per loop
         5        646 nsec per loop        686 nsec per loop
        10        838 nsec per loop        900 nsec per loop
        20       1.33 usec per loop       1.33 usec per loop
        25       1.52 usec per loop       1.55 usec per loop
        50       2.66 usec per loop       2.56 usec per loop
       100       5.13 usec per loop       4.74 usec per loop
       250       11.8 usec per loop       10.9 usec per loop
       500       25.5 usec per loop       24.1 usec per loop
      1000       51.6 usec per loop       49.9 usec per loop
    10_000        519 usec per loop        519 usec per loop
   100_000       6.88 msec per loop       6.46 msec per loop
 1_000_000       72.5 msec per loop       70.4 msec per loop
 5_000_000        364 msec per loop        359 msec per loop
10_000_000        735 msec per loop        719 msec per loop
50_000_000        3.67 sec per loop        3.72 sec per loop
```

My takeaways are:  
* For very short lists (fewer than ten elements), using a length hint is not helpful.  There is an additional cost of calling the length hint function, which seems to be around 50 ns.
    - By default, if a length hint is not available `list()` allocates enough memory for 8 elements.
    - if the `__length_hint__` method for the iterator were implemented in pure Python, this would presumably be a higher cost.
* The breakeven is iterables around ~25 elements.  
* For medium-length iterators (e.g. n = 50, 100, 250, 500), the implementation with `__length_hint__` is around 5% faster.

Next step will be pyperformance benchmarks.