@nmusolino 2019-06-29
Here's output for the pyperformance comparison, filtering on only changes labeled "Significant":

```$ pyperformance compare   ./build_97d15b1ee0/pyperformance_output_build_97d15b1ee0.json ./build_d6d90d1d3b/pyperformance_output_build_d6d90d1d3b.json  | grep -B3 'Significant'
```

Output shows two benchmarks slower and two faster.

```
### genshi_text ###
Mean +- std dev: 33.4 ms +- 0.4 ms -> 34.1 ms +- 1.1 ms: 1.02x slower
Significant (t=-4.67)
--

### richards ###
Mean +- std dev: 84.3 ms +- 1.2 ms -> 86.2 ms +- 1.6 ms: 1.02x slower
Significant (t=-7.29)
--

### sqlite_synth ###
Mean +- std dev: 3.15 us +- 0.04 us -> 3.05 us +- 0.05 us: 1.03x faster
Significant (t=11.38)
--

### unpickle_list ###
Mean +- std dev: 4.18 us +- 0.04 us -> 4.09 us +- 0.03 us: 1.02x faster
Significant (t=14.37)
```