@vstinner 2019-07-01
> For very short lists (fewer than ten elements), using a length hint is not helpful.

That's why the idea of using PyObject_LengthHint() has been abandoned. Instead, _PyObject_HasLen() was introduced to check if it's possible to get an object length without having to call a Python function (a Python function call is "slow" in such context):
* https://bugs.python.org/issue33234
* commit 372d705d958964289d762953d0a61622755f5386

```
commit 372d705d958964289d762953d0a61622755f5386
Author: Pablo Galindo <Pablogsal@gmail.com>
Date:   Sun Oct 28 20:16:26 2018 +0000

    bpo-33234 Improve list() pre-sizing for inputs with known lengths (GH-9846)
    
    The list() constructor isn't taking full advantage of known input
    lengths or length hints. This commit makes the constructor
    pre-size and not over-allocate when the input size is known (the
    input collection implements __len__). One on the main advantages is
    that this provides 12% difference in memory savings due to the difference
    between overallocating and allocating exactly the input size.
    
    For efficiency purposes and to avoid a performance regression for small
    generators and collections, the size of the input object is calculated using
    __len__ and not __length_hint__, as the later is considerably slower.
```

As I wrote previously, if you would like to explore the length hint path, you should experiment to add a new tp_length_hint slot to PyTypeObject (or maybe to PySequenceMethods and/or PyMappingMethods). The overhead of a Python function call is too high.