@iritkatriel 2022-04-11
https://bugs.python.org/issue26828 was rejected. Can this be closed?